refactor(arch): 封装Bilibili API客户端以完成三层架构
本次提交完成了对数据获取层的重构，将 `fetcher.py` 中的零散函数封装到一个专用的 `BilibiliClient` 类中。

主要变更：

- **创建 `bili_api_client.py`**:
  - 新建 `BilibiliClient` 类，通过依赖注入接收 `aiohttp.ClientSession`。
  - 所有原 `fetcher.py` 中的函数都被迁移为 `BilibiliClient` 的方法。公共方法（如 `get_all_user_videos`）和私有辅助方法（如 `_make_request`）被明确区分。

- **更新 `service.py`**:
  - `VideoService` 的依赖从 `aiohttp.ClientSession` 变更为 `BilibiliClient` 实例。
  - 现在通过调用 `self.client` 的方法来获取数据，而不是调用全局的 `fetch_*` 函数。这使得业务层与数据获取的实现细节完全解耦。

- **更新 `main.py`**:
  - 调整了应用的初始化顺序，现在是 `Session -> Client -> Service` 的清晰链条。
  - `main.py` 负责组装这些核心组件。

此次重构后，应用正式形成了经典的三层架构：
1.  **表示/协调层** (`main.py`)
2.  **业务逻辑层** (`service.py`)
3.  **数据访问层** (由 `bili_api_client.py` 和 `database.py` 组成)

这极大地提升了代码的模块化程度、可维护性和可测试性。
